+++ src/graph_onedrive/__init__.py
@@ -1,9 +1,3 @@
-__version__ = "0.0.1a10"
-
-# 1.2.0.dev1  # Development release
-# 1.2.0a1     # Alpha Release
-# 1.2.0b1     # Beta Release
-# 1.2.0rc1    # Release Candidate
-# 1.2.0       # Final Release
+__version__ = "0.1.0"
 
 from graph_onedrive._main import *